Move stream endpoint handler to st2stream package.
